+++ jest-puppeteer.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   exitOnPageError: false,
   launch: {
+    ignoreHTTPSErrors: true,
     headless: process.env.HEADLESS !== 'false',
     args: ['--start-maximized', '--no-sandbox'],
   },