+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.3.0 - September 8th, 2018
 - Fixed bugs in the HackerRank, URI Online Judge and Codeforces parsers
+- Added parser for UVa Online Judge and ICPC Live Archive
 
 # 2.2.4 - June 30th, 2018
 - Added support for AtCoder Beta